Fix Android JNI stats array length

+++ wgbridge-rs/src/jni_bindings.rs
@@ -337,7 +337,7 @@ pub extern "system" fn Java_net_kollnig_missioncontrol_wgbridge_Tunnel_nativeSta
                     stats.tun_write_failures_total,
                     stats.tun_write_failures_streak,
                 ];
-                match env.new_long_array(values.len() as i32) {
+                match env.new_long_array(values.len()) {
                     Ok(array) => {
                         let _ = array.set_region(env, 0, &values);
                         array.into_raw()
